requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
sqlalchemy_serializer==1.4.22
flask_wtf==1.2.2
gunicorn
flask_login